backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
sqlalchemy==2.0.35
alembic==1.13.2
aiosqlite==0.20.0
httpx==0.27.2
apscheduler==3.10.4
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1
cryptography==42.0.8
python-jose[cryptography]==3.3.0
python-multipart==0.0.12
